DataLik/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.3.2",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "plotly>=6.2.0",
    "pyyaml>=6.0.2",
    "streamlit-authenticator>=0.4.2",
    "streamlit>=1.47.1",
    "reportlab>=4.4.3",
    "sqlalchemy>=2.0.41",
    "psycopg2-binary>=2.9.10",
]
